Move scheme to root meta package to avoid import clutter
