+++ dataforge-data/src/commonMain/kotlin/hep/dataforge/data/DataFilter.kt
@@ -1,11 +1,6 @@
 package hep.dataforge.data
 
-import hep.dataforge.meta.Meta
-import hep.dataforge.meta.isEmpty
-import hep.dataforge.meta.scheme.Scheme
-import hep.dataforge.meta.scheme.SchemeSpec
-import hep.dataforge.meta.scheme.string
-import hep.dataforge.meta.scheme.wrap
+import hep.dataforge.meta.*
 import hep.dataforge.names.toName
