+++ dataforge-io/src/commonMain/kotlin/hep/dataforge/io/functions/RemoteFunctionServer.kt
@@ -8,7 +8,6 @@ import hep.dataforge.io.Responder
 import hep.dataforge.io.type
 import hep.dataforge.meta.get
 import hep.dataforge.meta.int
-import hep.dataforge.meta.scheme.int
 
 class RemoteFunctionServer(
     override val context: Context,
