+++ dataforge-meta/src/commonMain/kotlin/hep/dataforge/meta/mapMeta.kt
@@ -1,3 +1,5 @@
+@file:Suppress("UNUSED_PARAMETER")
+
 package hep.dataforge.meta
 
 import hep.dataforge.meta.descriptors.NodeDescriptor
