+++ dataforge-meta/src/commonTest/kotlin/hep/dataforge/meta/SchemeTest.kt
@@ -1,7 +1,5 @@
 package hep.dataforge.meta
 
-import hep.dataforge.meta.scheme.asScheme
-import hep.dataforge.meta.scheme.getProperty
 import kotlin.test.Test
 import kotlin.test.assertEquals
