+++ dataforge-meta/src/commonTest/kotlin/hep/dataforge/meta/SpecificationTest.kt
@@ -1,8 +1,5 @@
 package hep.dataforge.meta
 
-import hep.dataforge.meta.scheme.Scheme
-import hep.dataforge.meta.scheme.Specification
-import hep.dataforge.meta.scheme.numberList
 import hep.dataforge.names.Name
 import kotlin.test.Test
 import kotlin.test.assertEquals
