+++ dataforge-tables/src/commonMain/kotlin/hep/dataforge/tables/ColumnScheme.kt
@@ -1,9 +1,9 @@
 package hep.dataforge.tables
 
-import hep.dataforge.meta.scheme.Scheme
-import hep.dataforge.meta.scheme.SchemeSpec
-import hep.dataforge.meta.scheme.enum
-import hep.dataforge.meta.scheme.string
+import hep.dataforge.meta.Scheme
+import hep.dataforge.meta.SchemeSpec
+import hep.dataforge.meta.enum
+import hep.dataforge.meta.string
 import hep.dataforge.values.ValueType
 
 open class ColumnScheme : Scheme() {
